Vagrantfile: restart the containers unless-stopped
This change fixes the issue of stopped containers when the VMs are
restarted. The provisioning script, containing the `docker run` command,
is run only when the machine is provisioned for the first time. With
this change, the containers would automatically start when docker
starts.

Also, adds a note about how to start the containers manually.